requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
Pillow==10.0.1
pip==20.2.3
pymongo==3.11.4
pytz==2023.3.post1
setuptools==49.2.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==4.7.1
